SCRUM-140 feat: create custom hook
